[refactor] Moving전략 메소드명 변경 및 패키지 정리
