+++ src/main/java/racingcar/domain/car/CarStatus.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 public enum CarStatus {
     MOVE, STOP;
