+++ src/main/java/racingcar/domain/game/GameRoundResults.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.game;
 
 import java.util.ArrayList;
 import java.util.Collections;
