+++ src/main/java/racingcar/domain/strategy/CarMovingStrategy.java
@@ -1,13 +1,13 @@
 package racingcar.domain.strategy;
 
-import racingcar.domain.CarStatus;
+import racingcar.domain.car.CarStatus;
 
 public class CarMovingStrategy implements MovingStrategy{
 
     private static final int MOVING_STATUS_NUMBER = 4;
 
     @Override
-    public CarStatus playRound(int number) {
+    public CarStatus getMoveStatus(int number) {
         if (number >= MOVING_STATUS_NUMBER) {
             return CarStatus.MOVE;
         }
