+++ src/test/java/racingcar/application/RacingGameServiceTest.java
@@ -2,9 +2,9 @@ package racingcar.application;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.Cars;
-import racingcar.domain.GameResult;
+import racingcar.domain.car.Car;
+import racingcar.domain.car.Cars;
+import racingcar.domain.game.GameResult;
 
 import java.util.List;
