+++ src/test/java/racingcar/domain/car/CarsTest.java
@@ -1,10 +1,12 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import racingcar.domain.car.*;
+import racingcar.domain.game.GameRoundResult;
 import racingcar.domain.strategy.CarMovingStrategy;
 import racingcar.domain.strategy.RandomNumberGenerateStrategy;
